refactor: fix compile errors
This is broken code, but I want to see that we're writing to the
log files.
